pyproject.toml
[tool.poetry]
name = "image-workflows"
version = "0.1.2-dev0"
description = "Build and execute pipelines of polus plugins on Compute."
authors = ["Hamdah Shafqat Abbasi <hamdahshafqat.abbasi@nih.gv>"]
readme = "README.md"
packages = [{include = "image", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
typer = "^0.9.0"
pyyaml = "^6.0.1"
pydantic = "^2.6.1"
graphviz = "^0.20.3"
toil = "7.0.0"
sophios = "^0.1.4"
cwlref-runner = "1.0"
cwltool = "3.1.20240508115724"
polus-tools = { git = "https://github.com/PolusAI/tools.git" }



[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
nbconvert = "^7.11.0"
pytest = "^7.4.4"
bump2version = "^1.0.1"
pre-commit = "^3.3.3"
black = "^23.3.0"
ruff = "^0.0.274"
mypy = "^1.4.0"
pytest-xdist = "^3.3.1"
pytest-sugar = "^0.9.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]